@@ -1,14 +1,14 @@
 {
   "name": "malanna",
-  "version": "0.2.1",
+  "version": "0.3.0",
+  "license": "MIT",
+  "private": true,
   "author": {
     "name": "Nicholas Berlette",
     "email": "nick@berlette.com"
   },
-  "license": "MIT",
-  "private": true,
-  "description": "A small memorial website celebrating the life of a furry little friend of mine, Malanna. Powered by TypeScript, Next.js v12, Preact, and Prismic.io. Deployed on Vercel.",
   "repository": "nberlette/malanna",
+  "description": "A small memorial website celebrating the life of a furry little friend of mine, Malanna. Powered by TypeScript, Next.js v12, Preact, and Prismic.io. Deployed on Vercel.",
   "keywords": ["typescript", "next.js", "preact", "prismic", "ssg", "ssr", "malanna"],
   "bugs": "https://github.com/nberlette/malanna/issues",
   "homepage": "https://malanna.page",
@@ -24,30 +24,29 @@
   "dependencies": {
     "classnames": "2.3.1",
     "date-fns": "2.22.1",
-    "next": "12.0.3",
+    "next": "canary",
     "next-plugin-preact": "3.0.6",
-    "preact": "10.5.15",
-    "preact-render-to-string": "5.1.19",
-    "preact-ssr-prepass": "1.2.0",
+    "preact": "^10.5.15",
+    "preact-render-to-string": "^5.1.19",
+    "preact-ssr-prepass": "^1.2.0",
     "prismic-dom": "2.2.6",
     "prismic-javascript": "npm:@prismicio/client@5.1.0",
     "prismic-reactjs": "1.3.4",
-    "react": "npm:@preact/compat@17.0.3",
-    "react-dom": "npm:@preact/compat@17.0.3",
+    "react": "npm:@preact/compat@^17.0.2",
+    "react-dom": "npm:@preact/compat@^17.0.2",
     "react-ssr-prepass": "npm:preact-ssr-prepass@1.2.0",
-    "swr": "1.0.1"
+    "swr": "1.1.1"
   },
   "devDependencies": {
-    "@tailwindcss/typography": "0.4.1",
+    "@tailwindcss/typography": "0.5.0",
     "@types/node": "16.11.9",
     "@types/react": "17.0.35",
     "autoprefixer": "10.4.0",
     "eslint": "<8.0.0",
     "eslint-config-next": "12.0.3",
-    "postcss": "8.3.11",
+    "postcss": "8.4.5",
     "tailwindcss": "2.2.19",
-    "ts-node": "10.4.0",
     "tslib": "2.3.1",
-    "typescript": "4.4.4"
+    "typescript": "4.5.3"
   }
 }
